chore: introduce a CLI interface and a Makefile.
The CLI interface is not in use now.
