+++ .gitignore
@@ -13,3 +13,4 @@
 
 .idea
 .DS_Store
+apisix-ingress-controller
